refactor mic component in a separate
