+++ app/src/main/java/com/wearconnectivityexample/presentation/WearApp.kt
@@ -11,33 +11,11 @@ import androidx.compose.ui.tooling.preview.Devices
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.wear.compose.material.Button
 import androidx.wear.compose.material.Text
 import com.wearconnectivityexample.presentation.theme.ConnectivityAndroidExampleTheme
-import android.content.Intent
 import android.util.Log
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.padding
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
-import androidx.wear.compose.material.Chip
-import androidx.wear.compose.material.ChipDefaults
-import androidx.wear.compose.material.ExperimentalWearMaterialApi
-import androidx.wear.compose.material.ToggleButton
-import com.wearconnectivityexample.R
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.wrapContentSize
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
-import androidx.wear.compose.material.Icon
-import androidx.wear.compose.material.ToggleButton
-import androidx.wear.compose.material.ToggleButtonDefaults
 
 @Composable
 fun WearApp(currentCount: Int, increaseCount: () -> Unit, sendMessagesToClient: () -> Unit) {
@@ -70,31 +48,4 @@ fun Counter(count: Int) {
 @Composable
 fun DefaultPreview() {
     WearApp(0, { fun() {} }, { fun() {} })
-}
-
-
-@OptIn(ExperimentalWearMaterialApi::class)
-@Composable
-fun RecordVoiceScreen(onRecordClicked: () -> Unit) {
-    // Full-screen container with centered content.
-    Box(
-        modifier = Modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center
-    ) {
-        var checked by remember { mutableStateOf(true) }
-        ToggleButton(
-            checked = checked,
-            onCheckedChange = { checked = it },
-            enabled = true,
-        ) {
-            Icon(
-                painter = painterResource(id = R.drawable.ic_mic),
-                contentDescription = "airplane",
-                modifier =
-                Modifier
-                    .size(ToggleButtonDefaults.DefaultIconSize)
-                    .wrapContentSize(align = Alignment.Center),
-            )
-        }
-    }
 }
